[front/components/assistant_builder] - refactor: extract InstructionSuggestions into separate component
 - Moved instruction suggestion logic into a new `InstructionSuggestions` component for better modularity
 - Cleaned up imports by removing unused ones in `InstructionScreen.tsx`
 - Removed inline suggestion logic and replaced it with `InstructionSuggestions` component usage in `InstructionScreen.tsx`
